requirements.txt
# Core library for building multi-agent applications
pyautogen~=0.2.0

# LLM provider routing and management, used by AutoGen and configured directly
litellm~=1.34.0

# Used to load environment variables from the .env file (API keys, etc.)
python-dotenv~=1.0.0

# Required for the pre-processing step to read and extract text from PDF files
pypdf~=4.0.0

# Enables asynchronous file reading and writing operations
aiofiles~=23.0.0